pyproject.toml
[project]
name = "pptslide2video"
version = "0.1.0"
description = "Convert PPTX to animated MP4 with audio and subtitles"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "python-pptx>=0.6.23",
    "Jinja2>=3.1.0",
    "lxml>=5.0.0",
    "Pillow>=10.0.0",
    "comtypes>=1.4.0",
]

[project.scripts]
pptslide2video = "server:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"
